requirements.txt
cohere==4.34
python-dotenv==1.0.0
streamlit==1.28.1
weaviate_client==4.2b0
